Trying to get access to the V8 context by using the AppC generated code in build/generated. (Doesn't work.) In particular, we modified the CalcBootstrap.cpp to try to call example_initialize with the correct V8 context. This had multiple implications:
- The automated process tries to regenerate this script which we can't
allow. So we made it read-only. The build notes the error but keeps
going. Note that read-only doesn't prevent the file from getting
deleted, only modified.

- There is a dependency on example.c and the swig wrapper for it. We
dumped these directly into the generated directory. Initially we
about undefined symbols. But then it complains about multiple defined
symbols afterwards, so we commented one or both out. I think we need to
start with it in, then after the build fails, remove it without
cleaning to continue building. We also renamed example.c and added
extern "C" guards to .cpp
thinking maybe the build script wasn't compiling .c files, but I no
longer think that is the problem.

- We copied the build.xml from ~/Libary/Application
Support/Titatium/... and put it directly into the project so we could
modify it. The script tries to copy over the file from the temp
directory for some unknown reason (there seem to be no differences).
Since we made the file read only, the build process fails and aborts soawe removed that in the script.

- We tried two different functions in CalcBootstrap to call initialize.
Neither seemed to work. Josh says the namespace should get into
calc.swigex.My_Variable instead of swigex.My_Variable. Looking for both,
	we found neither.

- Josh suggested the global context which will only work from app.js,
	but when I tried it, it threw an exception and killed the app.

- I had to remove the CommonJS .js file I was using as a bootstrap to
make sure nothing was interfering with the process. It was a pain to
remove because the manifest needed to remove all entries related to it
(commonJS: false doesn't work).